Adjusted the 'checkCondition' routine so that it would return CHECKMAYBE.
CHECKMAYBE is returned when one or more conditions cannot be evaluated (there is no field to evaluate), but all other conditions are true.

if CHECKMAYBE is returned, a 'provisional' flag is set.  If a subsequent test returns CHECKTRUE then the system will check whether this is an 'exclusive' categorisation, and, if so, will only set the value if there is not an existing value.

The net effect should be that such values are only set where there is full information to set them.  This requirement originates in Allrisks categorisation of BB304140N, and BB304150N.   Categorisation of BB304140N needs county information which is not in the premiums, but categorisation of BB304150N requires commission information which is not in the risk.    Thus, in each case, the import with full information should be the defining import.

Note that, if both commission and county are required for a future import, we shall need some further thought!

git-svn-id: http://192.168.1.10/repos/toto@3342 b53939d4-f08c-4181-96d4-b201aa9c61c4