Removed the condition that an element only goes into a set as long as one of its parents are not already in the set.
This was originally designed to handle hierarchies Europe->London->Westminster - the attributes of Europe already apply to Westminster because London is in Europe.  But there is no harm in putting Westminster into Europe, and in other instances (e.g detecting duplicate customers where these are in a hierarchy) this caused problems.

It is noteworthy that, in the old regime, it was possible to add Westminster, then London, but not vice-versa!

git-svn-id: http://192.168.1.10/repos/toto@3992 b53939d4-f08c-4181-96d4-b201aa9c61c4